refactor: refactoring microphone
