+++ package.json
@@ -58,6 +58,7 @@
     "embla-carousel-react": "^8.6.0",
     "immer": "^11.1.3",
     "lucide-react": "^0.562.0",
+    "motion": "^12.31.1",
     "next": "16.1.1",
     "next-themes": "^0.4.6",
     "prom-client": "^15.1.3",
