+++ src/_pages/record/ui/LevelUpRecordPage.tsx
@@ -1,228 +1,39 @@
 'use client'
 
-import { useRouter, useSearchParams } from 'next/navigation'
-import { useCallback, useEffect, useState } from 'react'
-import { toast } from 'sonner'
-
-import { deleteCard } from '@/entities/card'
-import { useAccessToken } from '@/features/auth'
-import { startWarmup } from '@/features/levelup'
-import {
-  useCardDetails,
-  deleteAttempt,
-  FeedbackLoader,
-  FeedbackStatus,
-} from '@/features/levelup-feedback'
-import { MicrophoneBox, useMicrophone } from '@/features/record'
-import { completeAudioUpload, getAudioUrl, uploadAudio } from '@/features/record'
+import { FeedbackLoader } from '@/features/levelup-feedback'
+import { MicrophoneBox, useLevelUpRecordController } from '@/features/record'
 import { LevelUpHeader } from '@/shared'
 import { AlertModal, RecordTipBox, SubjectHeader } from '@/shared'
 import { Button } from '@/shared/ui/button'
 
 const RECORD_PROGRESS_VALUE = 100
 const RECORD_STEP_LABEL = '3/3'
-const REDIRECT_DELAY_MS = 1500
-type SupportedAudioContentType = 'audio/mp4' | 'audio/webm' | 'audio/wav' | 'audio/mpeg'
-const MIME_TYPE_DELIMITER = ';'
-const MIME_TO_CONTENT_TYPE_MAP: Record<string, SupportedAudioContentType> = {
-  'audio/webm': 'audio/webm',
-  'audio/mp4': 'audio/mp4',
-  'audio/wav': 'audio/wav',
-  'audio/mpeg': 'audio/mpeg',
-}
 export function LevelUpRecordPage() {
-  const router = useRouter()
-
-  const searchParams = useSearchParams()
-  const cardId = Number(searchParams.get('cardId'))
-  const attemptNo = Number(searchParams.get('attemptNo'))
-  const attemptId = Number(searchParams.get('attemptId'))
-
-  const accessToken = useAccessToken()
-
-  const { data } = useCardDetails(accessToken, cardId)
   const {
+    data,
+    isSubmittingFeedback,
+    uploadStatus,
+    isStartingWarmup,
+    warmupError,
     isMicAlertOpen,
-    setIsMicAlertOpen,
-    startRecording,
-    stopRecordingAndGetBlob,
-    recordedBlob,
+    isBackAlertOpen,
     isRecording,
     isPaused,
-    pauseRecording,
-    resumeRecording,
     elapsedSeconds,
-    getDurationSeconds,
-    clearRecordedBlob,
-    autoStopped,
-    resetAutoStopped,
-  } = useMicrophone()
-  const [isBackAlertOpen, setIsBackAlertOpen] = useState(false)
-  const [warmupError, setWarmupError] = useState(false)
-  const [isStartingWarmup, setIsStartingWarmup] = useState(false)
-  const [uploadStatus, setUploadStatus] = useState<FeedbackStatus | null>(null)
-  const [isSubmittingFeedback, setIsSubmittingFeedback] = useState(false)
-
-  useEffect(() => {
-    if (!warmupError) return
-
-    const timeoutId = window.setTimeout(() => {
-      router.push('/main')
-    }, REDIRECT_DELAY_MS)
-
-    return () => window.clearTimeout(timeoutId)
-  }, [router, warmupError])
-
-  const handleMicClick = async () => {
-    if (!accessToken || !cardId) return
-
-    if (isRecording) {
-      if (isPaused) {
-        resumeRecording()
-      } else {
-        pauseRecording()
-      }
-      return
-    }
-
-    setIsStartingWarmup(true)
-    const response = await startWarmup(accessToken, { cardId })
-    setIsStartingWarmup(false)
-
-    if (!response) {
-      setWarmupError(true)
-      toast.error('음성 녹음에 실패하였습니다. 잠시 후 다시 시도해주세요.')
-      return
-    }
-
-    const started = await startRecording()
-    if (!started) return
-  }
-
-  const handleBack = () => {
-    setIsBackAlertOpen(true)
-  }
-
-  const handleBackConfirm = async () => {
-    setIsBackAlertOpen(false)
-    if (accessToken && cardId && attemptId && attemptNo !== 1) {
-      await deleteAttempt(accessToken, cardId, attemptId)
-      router.push('/main')
-      return
-    }
-    if (accessToken && cardId) {
-      await deleteCard(accessToken, cardId)
-    }
-    router.push('/main')
-  }
-
-  const handleBackCancel = () => {
-    setIsBackAlertOpen(false)
-  }
-
-  const handleMicAlertClose = () => {
-    setIsMicAlertOpen(false)
-  }
-
-  const handleRecordingComplete = useCallback(async () => {
-    if (isSubmittingFeedback) return
-    if (!accessToken || !cardId) return
-
-    setIsSubmittingFeedback(true)
-    const completedBlob = await stopRecordingAndGetBlob()
-    if (!completedBlob) {
-      setIsSubmittingFeedback(false)
-      toast.error('녹음 파일을 생성하던 중 오류가 발생했습니다. 다시 녹음해주세요.')
-      return
-    }
-
-    const durationSeconds = getDurationSeconds()
-
-    const normalizedMimeType = completedBlob.type.split(MIME_TYPE_DELIMITER)[0]
-    const contentType = MIME_TO_CONTENT_TYPE_MAP[normalizedMimeType] ?? 'audio/webm'
-
-    const audioUrlResult = await getAudioUrl(accessToken, attemptId, contentType)
-    if (!audioUrlResult.ok) {
-      setIsSubmittingFeedback(false)
-      toast.error('오디오 업로드 URL을 가져오지 못했습니다. 다시 시도해주세요.')
-      return
-    }
-
-    const uploadUrl = audioUrlResult.data?.uploadUrl
-    const objectKey = audioUrlResult.data?.objectKey
-    const uploadContentType = audioUrlResult.data?.contentType
-
-    if (!uploadUrl || !objectKey || !uploadContentType) {
-      setIsSubmittingFeedback(false)
-      toast.error('오디오 업로드 정보를 가져오지 못했습니다.')
-      return
-    }
-
-    const uploadResult = await uploadAudio(uploadUrl, completedBlob, uploadContentType)
-    if (!uploadResult.ok) {
-      setIsSubmittingFeedback(false)
-      toast.error('오디오 업로드에 실패했습니다. 다시 시도해주세요.')
-      return
-    }
-
-    const completeResult = await completeAudioUpload(
-      accessToken,
-      cardId,
-      attemptId,
-      objectKey,
-      durationSeconds,
-    )
-    if (!completeResult.ok) {
-      setIsSubmittingFeedback(false)
-      toast.error('오디오 업로드에 실패했습니다. 다시 시도해주세요.')
-      return
-    }
-
-    if (completeResult.data?.status === 'PENDING') {
-      setUploadStatus('PENDING')
-    }
-
-    clearRecordedBlob()
-    const feedbackParams = new URLSearchParams({
-      cardId: String(cardId),
-      attemptId: String(attemptId),
-    })
-
-    if (attemptNo !== null) {
-      feedbackParams.set('attemptNo', String(attemptNo))
-    }
-
-    router.replace(`/levelup/feedback?${feedbackParams.toString()}`)
-  }, [
-    isSubmittingFeedback,
-    accessToken,
-    cardId,
-    attemptId,
-    attemptNo,
-    stopRecordingAndGetBlob,
-    getDurationSeconds,
-    clearRecordedBlob,
-    setIsSubmittingFeedback,
-    setUploadStatus,
-    router,
-  ])
-
-  useEffect(() => {
-    if (!autoStopped || !recordedBlob || isSubmittingFeedback) return
-
-    resetAutoStopped()
-    resetAutoStopped()
-    // defer to avoid calling setState synchronously inside the effect
-    setTimeout(() => {
-      void handleRecordingComplete()
-    }, 0)
-  }, [autoStopped, handleRecordingComplete, isSubmittingFeedback, recordedBlob, resetAutoStopped])
+    recordedBlob,
+    handleMicClick,
+    handleBackConfirm,
+    handleBackCancel,
+    handleMicAlertOpenChange,
+    handleBackAlertOpenChange,
+    handleRecordingComplete,
+  } = useLevelUpRecordController()
 
   return (
     <div className="flex h-full w-full flex-1 flex-col">
       <LevelUpHeader
         variant="recording"
-        onBack={handleBack}
+        onBack={() => handleBackAlertOpenChange(true)}
         progressValue={RECORD_PROGRESS_VALUE}
         stepLabel={RECORD_STEP_LABEL}
       />
@@ -259,7 +70,7 @@ export function LevelUpRecordPage() {
       </div>
       <AlertModal
         open={isBackAlertOpen}
-        onOpenChange={setIsBackAlertOpen}
+        onOpenChange={handleBackAlertOpenChange}
         title="학습을 취소하시겠습니까?"
         description="현재 생성한 카드가 삭제될 수 있습니다."
         action="나가기"
@@ -269,13 +80,13 @@ export function LevelUpRecordPage() {
       />
       <AlertModal
         open={isMicAlertOpen}
-        onOpenChange={setIsMicAlertOpen}
+        onOpenChange={handleMicAlertOpenChange}
         title="마이크 권한이 필요합니다."
         description="브라우저 설정에서 마이크 권한을 허용해주세요."
         action="확인"
         cancel="닫기"
-        onAction={handleMicAlertClose}
-        onCancel={handleMicAlertClose}
+        onAction={() => handleMicAlertOpenChange(false)}
+        onCancel={() => handleMicAlertOpenChange(false)}
       />
     </div>
   )
