+++ src/features/record/index.ts
@@ -1,6 +1,7 @@
 export { startMediaRecorder } from './model/startMediaRecorder'
 export { MicrophoneBox } from './ui/MicrophoneBox'
 export { useMicrophone } from './model/client/useMicrophone'
+export { useLevelUpRecordController } from './model/useLevelUpRecordController'
 export { getAudioUrl } from './api/getAudioUrl'
 export { uploadAudio } from './api/uploadAudio'
 export { completeAudioUpload } from './api/completeAudioUpload'
